Separate Node.js dependency (#265)
